requirements.txt
argparse>=1.4.0
mysql-connector-python>=8.0.22,<9.0.0
pulumi>=3.0.0,<4.0.0
pulumi-aws>=4.0.0,<5.0.0
pulumi-github>=4.0.0,<5.0.0
PyYAML>=5.4.1
boto3>=1.18.43
cryptography>=3.4.8